Separate fetch function from cli.py
